refactor(payment): Move v1 mappers to v1 folder
